[feature] Manage pg_hba.conf from Ansible
- In staging, one needs to use a shell script in /initdb.d/ because
the storage volume is not exposed (mounted) on the filesystem. On the
other hand, in production, edit the file directly (out of the same
template)
- Ensure that local (UNIX-domain socket) connections remain possible
without a password (`trust`)
- We now have a directory full of scripts for the staging
/initdb.d/, *and* a backup file bind-mounted read-only from elsewhere.
Convert the former into a directory mount, so that we don't have to
think of that each time. (It so happens that that volume needs to be
read-write, lest the sub-mount fails with “read-only filesystem”)
- Refactor to add a couple of variables while we are at it