[feature] Pass secrets (no matter their criticity) in configuration-as-code
Fixes
https://trello.com/c/Uqk6ma1h/20-g%C3%A9rer-servinggitlabjsonsecretsfile-dans-kubernetes

We need all secrets to at least be the same within any given serving
cluster. Additionally, there is a risk that a secret we think is
transient (e.g. to validate password reset links) is actually
not (e.g. for people who receive their email as an upgrade is in
progress). Better act safe, so all secrets in this commit come from
the production secrets on the gitlab.epfl.ch VM.

- Implement a `| to_omnibus` filter plugin to render the secrets in
the Ruby format expected by Omnibus
- When doing secrets in Kubernetes, source (with eval(File.read
"...")) another Ruby file that has been rendered from the EYAML
secrets using the aforementioned `| to_omnibus`
- Research for each secret whether it is precious, throwaway
or (currently) unused, and document that as-code.
- Document the fact that we might want to move to that discipline also
for the gitlab.epfl.ch VM