[feature] Manually provision PVC as well
This took a cussword-filled, absolute nightmare of an afternoon to
figure out. Bear with me.

Careful reading of
https://docs.gitlab.com/charts/charts/gitlab/gitaly/#git-repository-persistence
suggests that this is what we want when we have a manually provisioned
PV for the repositories (which we do). So far, so good.

No amount of careful reading will indicate exactly *where* the
`persistence:` dict should go in `values.yml`, though. Mind you, this
is standard suffering within the confines of Helm d(har|ra)ma. The correct
solution (which was none of the two redundant snippets in the previous
attempt) was determined by much cursing and trial and error.

And finally, (unless perhaps you already are a consummated master of
`StatefulSet.spec.volumeClaimTemplates`s, whose
[“documentation”](https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#volume-claim-templates)
is not exactly a beacon of clarity in and of itself) you will be
delighted to [learn
elsewhere](https://github.com/xinity/custom-gitlab/blob/master/doc/installation/storage.md#manually-creating-persistentvolumeclaims)
that the PVC name is *not* up for you to choose, either!

- Have PV and PVC point at each other by name, as per
https://kubernetes.io/docs/concepts/storage/persistent-volumes/#reserving-a-persistentvolume
- Prevent our PVC from falling into the default storage class
- Tell the Helm chart which PVC it will be using through labels, as
per ⌘F `matchLabels` after
https://docs.gitlab.com/charts/charts/gitlab/gitaly/#git-repository-persistence
- Sprout `gitlab_repo_volume_size` for a tiny bit of DRY