[fix] IPv6-related `CrashLoopBackoff` in `rke2-canal` pods
Before this change, the `kube-flannel` containers of pods in the
`rke2-canal` `DaemonSet` used to crash with the following message in
the logs:

```
Failed to find any valid interface to use: failed to get default v6
interface: Found default v6 route but could not determine interface
```

I take that as meaning that that `flanneld` Golang proggie was able to
figure out on its own which one of our VM's “physical” interfaces
bears the IPv4 default route; but then it couldn't be bothered to
connect the dots as far as IPv6 is concerned? 🤷‍♂️

[Ancient lore](https://github.com/flannel-io/flannel/issues/1573)
prophesized that The Next Release™ of Flannel would fix that, or
something close. Well, apparently it didn't.

- Help `flanneld` figure out which IPv6 interface is the main one, using
  the
  [`flannel.regexIface`](https://github.com/rancher/rke2-charts/blob/aae8652a0ef97291aeee5bc386136f64a0ffca6c/packages/rke2-canal/charts/values.yaml#L17)
  chart value, which
  [winds
  up](https://github.com/search?q=repo%3Arancher%2Frke2-charts%20FLANNELD_IFACE_REGEX&type=code)
  as the `FLANNELD_IFACE_REGEX` environment variable, which in turn
  [maps](https://github.com/flannel-io/flannel/blob/master/Documentation/configuration.md#environment-variables)
  to the
  [`-iface-regex`](https://github.com/search?q=repo%3Aflannel-io%2Fflannel%20iface-regex&type=code)
  command-line switch
- Merge that setting with whatever was already present in `rke2-canal`'s `HelmChartConfig`
  object, as set up at Rancher installation time